refactor: split auth_handler.go by responsibility
