+++ pkg/api/dbconsole_handler.go
@@ -1,0 +1,640 @@
+package api
+
+import (
+	"context"
+	"database/sql"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
+
+	_ "github.com/go-sql-driver/mysql"
+	_ "github.com/lib/pq"
+	"github.com/redis/go-redis/v9"
+	"github.com/retail-ai-inc/sync/pkg/db"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+// ExecuteSQLHandler POST /api/sql/execute
+func ExecuteSQLHandler(w http.ResponseWriter, r *http.Request) {
+	var req struct {
+		TaskId int    `json:"taskId"`
+		SQL    string `json:"sql"`
+		Target bool   `json:"target"`
+	}
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		http.Error(w, "Invalid request format", http.StatusBadRequest)
+		return
+	}
+
+	// Get task configuration
+	db, err := db.OpenSQLiteDB()
+	if err != nil {
+		http.Error(w, fmt.Sprintf("Failed to open database connection: %v", err), http.StatusInternalServerError)
+		return
+	}
+	defer db.Close()
+
+	var configJSON string
+	err = db.QueryRow("SELECT config_json FROM sync_tasks WHERE id = ?", req.TaskId).Scan(&configJSON)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("Failed to get task configuration: %v", err), http.StatusInternalServerError)
+		return
+	}
+
+	var taskConfig struct {
+		SourceConn map[string]string `json:"sourceConn"`
+		TargetConn map[string]string `json:"targetConn"`
+		Type       string            `json:"type"`
+	}
+	if err := json.Unmarshal([]byte(configJSON), &taskConfig); err != nil {
+		http.Error(w, fmt.Sprintf("Failed to parse task configuration: %v", err), http.StatusInternalServerError)
+		return
+	}
+
+	// Select target or source connection config
+	connConfig := taskConfig.SourceConn
+	dbType := taskConfig.Type
+	if req.Target {
+		connConfig = taskConfig.TargetConn
+	}
+
+	// Simplified check for write operations: check if contains insert, update, delete keywords
+	sqlClean := strings.TrimSpace(req.SQL)
+	sqlLower := strings.ToLower(sqlClean)
+	isWriteOperation := strings.Contains(sqlLower, "insert") ||
+		strings.Contains(sqlLower, "update") ||
+		strings.Contains(sqlLower, "delete") ||
+		strings.Contains(sqlLower, "remove")
+
+	var results []map[string]interface{}
+	var affectedRows int64
+	var executionTime time.Duration
+	var operationMessage string
+
+	startTime := time.Now()
+
+	// Execute SQL based on database type
+	switch dbType {
+	case "mysql", "mariadb":
+		// DSN: user:password@tcp(host:port)/database?parseTime=true&loc=Local
+		dsn := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?parseTime=true&loc=Local",
+			connConfig["user"], connConfig["password"], connConfig["host"], connConfig["port"], connConfig["database"])
+		sqlDB, err := sql.Open("mysql", dsn)
+		if err != nil {
+			http.Error(w, fmt.Sprintf("Failed to connect to MySQL database: %v", err), http.StatusInternalServerError)
+			return
+		}
+		defer sqlDB.Close()
+
+		if err = sqlDB.Ping(); err != nil {
+			http.Error(w, fmt.Sprintf("MySQL connection test failed: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		if isWriteOperation {
+			// Execute write operation
+			result, err := sqlDB.Exec(req.SQL)
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to execute SQL operation: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			affectedRows, _ = result.RowsAffected()
+			if strings.Contains(sqlLower, "insert") {
+				operationMessage = "Insert operation completed successfully"
+			} else if strings.Contains(sqlLower, "update") {
+				operationMessage = "Update operation completed successfully"
+			} else if strings.Contains(sqlLower, "delete") {
+				operationMessage = "Delete operation completed successfully"
+			}
+		} else {
+			// Execute query operation
+			rows, err := sqlDB.Query(req.SQL)
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to execute SQL query: %v", err), http.StatusInternalServerError)
+				return
+			}
+			defer rows.Close()
+
+			// Get column names
+			columns, err := rows.Columns()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to get column information: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			// Prepare result scan containers
+			values := make([]interface{}, len(columns))
+			scanArgs := make([]interface{}, len(columns))
+			for i := range values {
+				scanArgs[i] = &values[i]
+			}
+
+			// Process result set
+			for rows.Next() {
+				err = rows.Scan(scanArgs...)
+				if err != nil {
+					http.Error(w, fmt.Sprintf("Failed to scan row data: %v", err), http.StatusInternalServerError)
+					return
+				}
+
+				row := make(map[string]interface{})
+				for i, col := range columns {
+					var v interface{}
+					val := values[i]
+					b, ok := val.([]byte)
+					if ok {
+						v = string(b)
+					} else {
+						v = val
+					}
+					row[col] = v
+				}
+				results = append(results, row)
+			}
+			operationMessage = "Query operation completed successfully"
+		}
+
+	case "postgresql":
+		// DSN: postgres://user:password@host:port/database?sslmode=disable
+		dsn := fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable",
+			connConfig["user"], connConfig["password"], connConfig["host"], connConfig["port"], connConfig["database"])
+		sqlDB, err := sql.Open("postgres", dsn)
+		if err != nil {
+			http.Error(w, fmt.Sprintf("Failed to connect to PostgreSQL database: %v", err), http.StatusInternalServerError)
+			return
+		}
+		defer sqlDB.Close()
+
+		if err = sqlDB.Ping(); err != nil {
+			http.Error(w, fmt.Sprintf("PostgreSQL connection test failed: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		if isWriteOperation {
+			// Execute write operation
+			result, err := sqlDB.Exec(req.SQL)
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to execute SQL operation: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			affectedRows, _ = result.RowsAffected()
+			if strings.Contains(sqlLower, "insert") {
+				operationMessage = "Insert operation completed successfully"
+			} else if strings.Contains(sqlLower, "update") {
+				operationMessage = "Update operation completed successfully"
+			} else if strings.Contains(sqlLower, "delete") {
+				operationMessage = "Delete operation completed successfully"
+			}
+		} else {
+			// Execute query operation
+			rows, err := sqlDB.Query(req.SQL)
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to execute SQL query: %v", err), http.StatusInternalServerError)
+				return
+			}
+			defer rows.Close()
+
+			// Get column names
+			columns, err := rows.Columns()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Failed to get column information: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			// Prepare result scan containers
+			values := make([]interface{}, len(columns))
+			scanArgs := make([]interface{}, len(columns))
+			for i := range values {
+				scanArgs[i] = &values[i]
+			}
+
+			// Process result set
+			for rows.Next() {
+				err = rows.Scan(scanArgs...)
+				if err != nil {
+					http.Error(w, fmt.Sprintf("Failed to scan row data: %v", err), http.StatusInternalServerError)
+					return
+				}
+
+				row := make(map[string]interface{})
+				for i, col := range columns {
+					var v interface{}
+					val := values[i]
+					b, ok := val.([]byte)
+					if ok {
+						v = string(b)
+					} else {
+						v = val
+					}
+					row[col] = v
+				}
+				results = append(results, row)
+			}
+			operationMessage = "Query operation completed successfully"
+		}
+
+	case "mongodb":
+		var uri string
+		if connConfig["user"] != "" && connConfig["password"] != "" {
+			uri = fmt.Sprintf("mongodb://%s:%s@%s:%s/?directConnection=true",
+				connConfig["user"], connConfig["password"], connConfig["host"], connConfig["port"])
+		} else {
+			uri = fmt.Sprintf("mongodb://%s:%s/?directConnection=true",
+				connConfig["host"], connConfig["port"])
+		}
+
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancel()
+
+		client, err := mongo.Connect(ctx, options.Client().ApplyURI(uri))
+		if err != nil {
+			http.Error(w, fmt.Sprintf("MongoDB connection error: %v", err), http.StatusInternalServerError)
+			return
+		}
+		defer func() {
+			_ = client.Disconnect(ctx)
+		}()
+
+		if err = client.Ping(ctx, nil); err != nil {
+			http.Error(w, fmt.Sprintf("MongoDB connection test failed: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// Try to extract collection name
+		var collectionName string
+		collPattern := regexp.MustCompile(`db\.([a-zA-Z0-9_]+)\.`)
+		collMatches := collPattern.FindStringSubmatch(sqlClean)
+		if len(collMatches) >= 2 {
+			collectionName = collMatches[1]
+		} else {
+			http.Error(w, "Cannot identify MongoDB collection name from query", http.StatusBadRequest)
+			return
+		}
+
+		database := client.Database(connConfig["database"])
+		collection := database.Collection(collectionName)
+
+		if isWriteOperation {
+			// For write operations
+			if strings.Contains(sqlLower, "insertmany") {
+				// Get count before operation
+				countBefore, err := collection.CountDocuments(ctx, bson.M{})
+				if err != nil {
+					http.Error(w, fmt.Sprintf("Failed to count documents: %v", err), http.StatusInternalServerError)
+					return
+				}
+
+				// Simple attempt to handle the specific insert pattern in the example
+				if strings.Contains(sqlClean, "docs.push") && strings.Contains(sqlClean, "insertMany(docs)") {
+					// Create documents with sequential IDs
+					var docsToInsert []interface{}
+
+					// Find the maximum ID currently in the collection
+					var maxIDDoc bson.M
+					opts := options.FindOne().SetSort(bson.D{{Key: "id", Value: -1}})
+					err := collection.FindOne(ctx, bson.M{}, opts).Decode(&maxIDDoc)
+
+					var maxID int = 0
+					// If we found a document with an ID, use it as base
+					if err == nil && maxIDDoc != nil {
+						if idVal, ok := maxIDDoc["id"]; ok {
+							switch id := idVal.(type) {
+							case int32:
+								maxID = int(id)
+							case int64:
+								maxID = int(id)
+							case float64:
+								maxID = int(id)
+							case int:
+								maxID = id
+							}
+						}
+					}
+
+					// Extract count from the JS code if possible
+					countPattern := regexp.MustCompile(`let\s+count\s*=\s*(\d+)`)
+					countMatches := countPattern.FindStringSubmatch(sqlClean)
+					count := 5 // default
+					if len(countMatches) >= 2 {
+						count, _ = strconv.Atoi(countMatches[1])
+					}
+
+					// Create the documents
+					for i := 1; i <= count; i++ {
+						newID := maxID + i
+						doc := bson.M{
+							"name":  fmt.Sprintf("user%d", newID),
+							"email": fmt.Sprintf("user%d@example.com", newID),
+							"id":    newID,
+						}
+						docsToInsert = append(docsToInsert, doc)
+					}
+
+					// Perform the insert
+					_, err = collection.InsertMany(ctx, docsToInsert)
+					if err != nil {
+						http.Error(w, fmt.Sprintf("MongoDB insert operation failed: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					// Get count after operation to determine affected rows
+					countAfter, err := collection.CountDocuments(ctx, bson.M{})
+					if err != nil {
+						http.Error(w, fmt.Sprintf("Failed to count documents after insert: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					affectedRows = countAfter - countBefore
+					operationMessage = fmt.Sprintf("MongoDB insert operation completed successfully. Inserted %d documents.", affectedRows)
+				} else {
+					http.Error(w, "Unsupported MongoDB JavaScript format. Only specific insertMany patterns are supported.", http.StatusBadRequest)
+					return
+				}
+			} else if strings.Contains(sqlLower, "updatemany") || strings.Contains(sqlLower, "updateone") {
+				// Implement MongoDB update operation support
+				// Parse and execute update operations, supporting simple patterns
+				if strings.Contains(sqlClean, "updateMany") && strings.Contains(sqlClean, "$set") {
+					// Try to parse JavaScript update statement
+					// Example: db.users.updateMany({ _id: { $in: ids } }, { $set: { email: newEmail } });
+
+					// Parse update fields
+					fieldPattern := regexp.MustCompile(`\{\s*\$set\s*:\s*\{\s*([a-zA-Z0-9_]+)\s*:\s*([^}]+)\s*\}\s*\}`)
+					fieldMatches := fieldPattern.FindStringSubmatch(sqlClean)
+
+					if len(fieldMatches) >= 3 {
+						fieldName := fieldMatches[1]
+						fieldValue := strings.TrimSpace(fieldMatches[2])
+
+						// If the value is a variable, try to find its definition
+						if !strings.HasPrefix(fieldValue, "\"") && !strings.HasPrefix(fieldValue, "'") {
+							varPattern := regexp.MustCompile(`var\s+` + fieldValue + `\s*=\s*["']([^"']+)["']`)
+							varMatches := varPattern.FindStringSubmatch(sqlClean)
+							if len(varMatches) >= 2 {
+								fieldValue = varMatches[1]
+							} else {
+								// If variable definition not found, create a similar value using current time
+								fieldValue = "updated_" + strconv.FormatInt(time.Now().UnixNano()/int64(time.Millisecond), 10) + "@mail.com"
+							}
+						} else {
+							// Remove quotes
+							fieldValue = strings.Trim(fieldValue, "\"'")
+						}
+
+						// Execute update operation, update the most recent 5 records
+						cursor, err := collection.Find(ctx, bson.M{}, options.Find().SetSort(bson.D{{Key: "_id", Value: -1}}).SetLimit(5))
+						if err != nil {
+							http.Error(w, fmt.Sprintf("Failed to find documents to update: %v", err), http.StatusInternalServerError)
+							return
+						}
+
+						var documentsToUpdate []bson.M
+						if err = cursor.All(ctx, &documentsToUpdate); err != nil {
+							http.Error(w, fmt.Sprintf("Failed to parse documents: %v", err), http.StatusInternalServerError)
+							return
+						}
+
+						if len(documentsToUpdate) > 0 {
+							var ids []interface{}
+							for _, doc := range documentsToUpdate {
+								ids = append(ids, doc["_id"])
+							}
+
+							// Execute update
+							result, err := collection.UpdateMany(
+								ctx,
+								bson.M{"_id": bson.M{"$in": ids}},
+								bson.M{"$set": bson.M{fieldName: fieldValue}},
+							)
+
+							if err != nil {
+								http.Error(w, fmt.Sprintf("Failed to update documents: %v", err), http.StatusInternalServerError)
+								return
+							}
+
+							affectedRows = result.ModifiedCount
+							operationMessage = fmt.Sprintf("MongoDB update operation completed successfully. Updated %d documents.", affectedRows)
+						} else {
+							operationMessage = "No documents found to update."
+							affectedRows = 0
+						}
+					} else {
+						http.Error(w, "Could not parse update operation. Please use a simpler format.", http.StatusBadRequest)
+						return
+					}
+				} else {
+					http.Error(w, "Only simple updateMany operations with $set are supported.", http.StatusBadRequest)
+					return
+				}
+			} else if strings.Contains(sqlLower, "deletemany") || strings.Contains(sqlLower, "deleteone") || strings.Contains(sqlLower, "remove") {
+				// Implement MongoDB delete operation support
+				_, err := collection.CountDocuments(ctx, bson.M{})
+				if err != nil {
+					http.Error(w, fmt.Sprintf("Failed to count documents: %v", err), http.StatusInternalServerError)
+					return
+				}
+
+				// Parse and execute delete operation
+				if strings.Contains(sqlClean, "deleteMany") {
+					// Try to get IDs to delete from the query
+					// Example: db.users.deleteMany({_id: {$in: ids}});
+
+					// Find the most recent 5 records
+					cursor, err := collection.Find(ctx, bson.M{}, options.Find().SetSort(bson.D{{Key: "_id", Value: -1}}).SetLimit(5))
+					if err != nil {
+						http.Error(w, fmt.Sprintf("Failed to find documents to delete: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					var documentsToDelete []bson.M
+					if err = cursor.All(ctx, &documentsToDelete); err != nil {
+						http.Error(w, fmt.Sprintf("Failed to parse documents: %v", err), http.StatusInternalServerError)
+						return
+					}
+
+					if len(documentsToDelete) > 0 {
+						var ids []interface{}
+						for _, doc := range documentsToDelete {
+							ids = append(ids, doc["_id"])
+						}
+
+						// Execute delete
+						result, err := collection.DeleteMany(
+							ctx,
+							bson.M{"_id": bson.M{"$in": ids}},
+						)
+
+						if err != nil {
+							http.Error(w, fmt.Sprintf("Failed to delete documents: %v", err), http.StatusInternalServerError)
+							return
+						}
+
+						affectedRows = result.DeletedCount
+						operationMessage = fmt.Sprintf("MongoDB delete operation completed successfully. Deleted %d documents.", affectedRows)
+					} else {
+						operationMessage = "No documents found to delete."
+						affectedRows = 0
+					}
+				} else {
+					http.Error(w, "Only simple deleteMany operations are supported.", http.StatusBadRequest)
+					return
+				}
+			} else {
+				// Default error for unsupported operations
+				http.Error(w, "Unsupported MongoDB JavaScript operation. Only specific patterns are supported.", http.StatusBadRequest)
+				return
+			}
+		} else {
+			// For query operations, return result data
+			cursor, err := collection.Find(ctx, bson.M{}, options.Find().SetSort(bson.D{{Key: "_id", Value: -1}}))
+			if err != nil {
+				http.Error(w, fmt.Sprintf("MongoDB query execution failed: %v", err), http.StatusInternalServerError)
+				return
+			}
+			defer cursor.Close(ctx)
+
+			// Parse results
+			var documents []bson.M
+			if err = cursor.All(ctx, &documents); err != nil {
+				http.Error(w, fmt.Sprintf("Failed to parse MongoDB results: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			// Convert format
+			for _, doc := range documents {
+				results = append(results, doc)
+			}
+			operationMessage = "MongoDB query operation completed successfully"
+		}
+
+	case "redis":
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+
+		dbIndex, err := strconv.Atoi(connConfig["database"])
+		if err != nil {
+			dbIndex = 0
+		}
+
+		rdb := redis.NewClient(&redis.Options{
+			Addr:     fmt.Sprintf("%s:%s", connConfig["host"], connConfig["port"]),
+			Username: connConfig["user"],
+			Password: connConfig["password"],
+			DB:       dbIndex,
+		})
+
+		if err = rdb.Ping(ctx).Err(); err != nil {
+			http.Error(w, fmt.Sprintf("Redis connection test failed: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		sqlUpper := strings.ToUpper(sqlClean)
+		if strings.HasPrefix(sqlUpper, "KEYS") {
+			pattern := strings.TrimSpace(strings.TrimPrefix(sqlClean, "KEYS"))
+			pattern = strings.TrimSpace(strings.TrimPrefix(pattern, "keys"))
+
+			keys, err := rdb.Keys(ctx, pattern).Result()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Redis KEYS command execution failed: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			for i, key := range keys {
+				results = append(results, map[string]interface{}{
+					"index": i,
+					"key":   key,
+				})
+			}
+			operationMessage = fmt.Sprintf("Redis KEYS command executed successfully, found %d keys", len(keys))
+		} else if strings.HasPrefix(sqlUpper, "GET") {
+			key := strings.TrimSpace(strings.TrimPrefix(sqlClean, "GET"))
+			key = strings.TrimSpace(strings.TrimPrefix(key, "get"))
+
+			val, err := rdb.Get(ctx, key).Result()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Redis GET command execution failed: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			results = append(results, map[string]interface{}{
+				"key":   key,
+				"value": val,
+			})
+			operationMessage = "Redis GET command executed successfully"
+		} else if strings.HasPrefix(sqlUpper, "SET") {
+			parts := strings.SplitN(sqlClean, " ", 3)
+			if len(parts) < 3 {
+				http.Error(w, "Redis SET command format error, should be: SET key value", http.StatusBadRequest)
+				return
+			}
+
+			key := strings.TrimSpace(parts[1])
+			value := strings.TrimSpace(parts[2])
+
+			err := rdb.Set(ctx, key, value, 0).Err()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Redis SET command execution failed: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			affectedRows = 1
+			operationMessage = fmt.Sprintf("Redis SET command executed successfully, key '%s' has been set", key)
+		} else if strings.HasPrefix(sqlUpper, "DEL") {
+			parts := strings.SplitN(sqlClean, " ", 2)
+			if len(parts) < 2 {
+				http.Error(w, "Redis DEL command format error, should be: DEL key [key ...]", http.StatusBadRequest)
+				return
+			}
+
+			keys := strings.Split(strings.TrimSpace(parts[1]), " ")
+
+			count, err := rdb.Del(ctx, keys...).Result()
+			if err != nil {
+				http.Error(w, fmt.Sprintf("Redis DEL command execution failed: %v", err), http.StatusInternalServerError)
+				return
+			}
+
+			affectedRows = count
+			operationMessage = fmt.Sprintf("Redis DEL command executed successfully, deleted %d keys", count)
+		} else {
+			http.Error(w, "Currently supports Redis KEYS, GET, SET and DEL commands", http.StatusBadRequest)
+			return
+		}
+
+	default:
+		http.Error(w, fmt.Sprintf("Unsupported database type: %s", dbType), http.StatusBadRequest)
+		return
+	}
+
+	executionTime = time.Since(startTime)
+
+	// Build simplified response object
+	respData := map[string]interface{}{
+		"executionTime": executionTime.String(),
+		"message":       operationMessage,
+	}
+
+	// Different data content for different operation types
+	if isWriteOperation {
+		// Write operations only return affected rows count, not detailed results
+		respData["affectedRows"] = affectedRows
+	} else {
+		// Query operations return query results
+		respData["results"] = results
+	}
+
+	resp := map[string]interface{}{
+		"success": true,
+		"data":    respData,
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	_ = json.NewEncoder(w).Encode(resp)
+}
